rbd-target-gw: done crash on Config failures
The halt/clearconfig path assmes the Config object is setup
but if the network is down or the ceph cluster is unreachable
Config initialization will fail. We then do not want to try and
run clearconfig() because functions like drop_target and drop_maps
want to try and access the config's config object which might not
be even set.

This just has __main__ exit on Config creation failures since
there is nothing to cleanup/unwind.