Force rtslib bs cache to be updated
If gwcli or ceph-ansible creates/deletes a disk then the rtslib bs cache
will be stale. This adds a invalidate caches call to force the cache to
be updated during device creation. For object creations done via
configfs scanning (querys via storage_objects/storage_object refs) this
rtslib patch

https://github.com/open-iscsi/rtslib-fb/pull/141

bypasses the cache.

Note that with that patch we could also manage the index creation and
completely avoid the cache, but it looks like it is a gluster and
targetcli specific issue that causes dev creation to take a minute for
one device. I tested 1024 devices and it was less than a second for each
device.