Fix alua group creation failure handling
If rbd-target-gw crashes we can leave the storage objects and its alua
groups setup in the kernel. On restart we will then crash because there
is no tpg_id on the alua_group. This seems to be leftover from a very
old alua class that has since moved to rtslib and dropped that field.

This patch has us remove the old alua group and recreate it in case
there are new default settings that need to be used.